requirements.txt
python-telegram-bot==13.7
gspread==5.1.1
oauth2client==4.1.3
python-dotenv==0.19.2
standard-imghdr==3.13.0
urllib3==1.26.15